cmd/observe: move flows specific code to flows*.go files
Follow-up for commit 700d6bcba89b ("cmd/observe: rename flow-related
functions and types"), separated to ease review.

This separates the generic observe functionality from the functionality
specific to flows. This follows the file naming scheme used for agent
and debug events.

No functional changes. This is preparatory refactoring for the
introduction of the `hubble observe flows` subcommand.
